+++ private/pkg/manifest/manifest.go
@@ -12,32 +12,33 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// A manifest is a file containing a list of paths and their hash digests,
-// canonically ordered by path in increasing lexicographical order. Manifests
-// are encoded as:
+// Package manifest defines generic content-addressable storage APIs which are
+// used to store and verify Buf modules. The primary data types relevant to
+// consumers of this package are [Manifest], [Blob], [BlobSet], and [Digest].
 //
-//	<digest type>:<digest>[SP][SP]<path>[LF]
-//
-// "shake256" is the only supported digest type. The digest is 64 bytes of hex
-// encoded output of SHAKE256. See golang.org/x/crypto/sha3 and FIPS 202 for
-// details on the SHAKE hash.
-//
-// [Manifest] can read and write manifest files. Canonical form is produced
-// when serialized ([Manifest.MarshalText]). Non-canonical form is a valid
-// manifest and will not produce errors when deserializing.
-//
+// [Manifest] can read and write manifest files. The canonical form of a
+// manifest is produced when serialized with [Manifest.MarshalText].
 // Interacting with a manifest is typically by path ([Manifest.Paths],
-// [Manifest.DigestFor]) or by a [Digest] ([Manifest.PathsFor]).
+// [Manifest.DigestFor]) or by a [Digest] ([Manifest.PathsFor]). Note that it
+// is possible for multiple paths in the manifest to have to same digest (and
+// content), however a given path only has one digest.
 //
-// [Blob] represents file content and its digest. [BlobSet] collects related
-// blobs together into a set. [NewMemoryBlob] provides an in-memory
-// implementation. A manifest, being a file, is also a blob ([Manifest.Blob]).
+// [Blob] represents file content and its digest. [NewMemoryBlob] provides an
+// in-memory implementation of a Blob. A manifest, being a file, is also a blob.
+// The [Manifest.Blob] function converts a manifest to a Blob.
+//
+// [BlobSet] collects related blobs together into a unique set (de-duplicating
+// blobs with the same digest and content).
 //
 // Blobs are anonymous files and a manifest gives names to anonymous files.
 // It's possible to view a manifest and its associated blobs as a file system.
-// [NewBucket] creates a storage bucket from a manifest and blob set.
-// [NewFromBucket] does the inverse: the creation of a manifest and blob set
-// from a storage bucket.
+// [NewFromBucket] creates a manifest and blob set from a storage bucket.
+//
+// Aside from the [Manifest.MarshalText] encoding, this package does not define
+// the representation of its data types on the wire. See the
+// private/bufpkg/bufmanifest package and
+// buf/alpha/registry/v1alpha1/module.proto for details on how these data types
+// are represented in Protobuf.
 package manifest
 
 import (
@@ -99,7 +100,7 @@ func NewFromReader(manifest io.Reader) (*Manifest, error) {
 		}
 	}
 	if err := scanner.Err(); err != nil {
-		if err == errNoFinalNewline {
+		if errors.Is(err, errNoFinalNewline) {
 			return nil, newError(lineno, "partial record")
 		}
 		return nil, err
@@ -185,8 +186,7 @@ func (m *Manifest) Paths() []string {
 	return pathsCopy
 }
 
-// Digests returns all unique digests in the manifest.
-// Order is by insertion order.
+// Digests returns all unique digests in the manifest, in insertion order.
 func (m *Manifest) Digests() []Digest {
 	digests := make([]Digest, 0, len(m.digestToPaths))
 	addedDigests := make(map[string]struct{}, len(m.digestToPaths))
@@ -228,7 +228,7 @@ func (m *Manifest) Range(f func(path string, digest Digest) error) error {
 
 // PathsFor returns one or more matching path for a given digest. The digest is
 // expected to be a lower-case hex encoded value. Returned paths are ordered by insertion time.
-// Paths is nil and ok is false if no paths are found.
+// Returns (nil, false) if no paths are found.
 func (m *Manifest) PathsFor(digest string) ([]string, bool) {
 	paths, ok := m.digestToPaths[digest]
 	if !ok || len(paths) == 0 {
@@ -238,7 +238,7 @@ func (m *Manifest) PathsFor(digest string) ([]string, bool) {
 }
 
 // DigestFor returns the matching digest for the given path. The path must be an
-// exact match. Digest is nil and ok is false if no digest is found.
+// exact match. Returns (nil, false) if no digest is found.
 func (m *Manifest) DigestFor(path string) (*Digest, bool) {
 	digest, ok := m.pathToDigest[path]
 	if !ok {
@@ -247,7 +247,21 @@ func (m *Manifest) DigestFor(path string) (*Digest, bool) {
 	return &digest, true
 }
 
-// MarshalText encodes the manifest into its canonical form.
+// MarshalText encodes the manifest into its canonical form, consisting of
+// an ordered list of paths and their hash digests. Manifests are encoded as:
+//
+//	<digest_type>:<digest>[SP][SP]<path>[LF]
+//
+// The only supported digest_type is shake256. The digest is 64 bytes of hex
+// encoded output of SHAKE256. See golang.org/x/crypto/sha3 and FIPS 202 for
+// details on the SHAKE hash.
+//
+// An example encoded manifest for the acme/petapis module is:
+//
+//	shake256:cd22db48cf7c274bbffcb5494a854000cd21b074df7c6edabbd0102c4be8d7623e3931560fcda7acfab286ae1d4f506911daa31f223ee159f59ffce0c7acbbaa  buf.lock
+//	shake256:3b353aa5aacd11015e8577f16e2c4e7a242ce773d8e3a16806795bb94f76e601b0db9bf42d5e1907fda63303e1fa1c65f1c175ecc025a3ef29c3456ad237ad84  buf.md
+//	shake256:7c88a20cf931702d042a4ddee3fde5de84814544411f1c62dbf435b1b81a12a8866a070baabcf8b5a0d31675af361ccb2d93ddada4cdcc11bab7ea3d8d7c4667  buf.yaml
+//	shake256:9db25155eafd19b36882cff129daac575baa67ee44d1cb1fd3894342b28c72b83eb21aa595b806e9cb5344759bc8308200c5af98e4329aa83014dde99afa903a  pet/v1/pet.proto
 func (m *Manifest) MarshalText() ([]byte, error) {
 	var coded bytes.Buffer
 	paths := m.Paths()
@@ -261,7 +275,8 @@ func (m *Manifest) MarshalText() ([]byte, error) {
 	return coded.Bytes(), nil
 }
 
-// UnmarshalText decodes a manifest from text.
+// UnmarshalText decodes a manifest from text. See [Manifest.MarshalText] for
+// the expected encoding of a manifest.
 //
 // See [NewFromReader] if your manifest is available in an io.Reader.
 func (m *Manifest) UnmarshalText(text []byte) error {
@@ -275,7 +290,9 @@ func (m *Manifest) UnmarshalText(text []byte) error {
 	return nil
 }
 
-// Blob returns the manifest as a blob.
+// Blob returns the manifest as a blob. The Blob content is set to the
+// canonical representation of a manifest ([Manifest.MarshalText]),
+// and the digest is set to the SHAKE256 digest of the content.
 func (m *Manifest) Blob() (Blob, error) {
 	manifestText, err := m.MarshalText()
 	if err != nil {